Move unit conversions to new module
